requirements.txt
aiosqlite==0.21.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.13
fastapi-mcp==0.3.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.10.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
-e git+https://github.com/j4mesmorais/financeiro.git@794e9be3aa54ee0ecfeaebab09fe2b0028db1f02#egg=UNKNOWN
urllib3==2.5.0
uvicorn==0.34.3
